executor/requirements.txt
influxdb-client
pika
paho-mqtt
requests